ARM7: Fix regression with Thumb branches

+++ src/arm/isa-thumb.c
@@ -308,8 +308,8 @@ DEFINE_LOAD_STORE_MULTIPLE_THUMB(STMIA,
 #define DEFINE_CONDITIONAL_BRANCH_THUMB(COND) \
 	DEFINE_INSTRUCTION_THUMB(B ## COND, \
 		if (ARM_COND_ ## COND) { \
-			uint8_t immediate = opcode; \
-			cpu->gprs[ARM_PC] += immediate << 1; \
+			int8_t immediate = opcode; \
+			cpu->gprs[ARM_PC] += (int32_t) immediate << 1; \
 			THUMB_WRITE_PC; \
 		})
